added not fount on lists services

+++ src/services/list/delete.ts
@@ -1,3 +1,4 @@
+import { NotFoundError } from "@presentation";
 import { ClientTypesAdapter, IClientRepository, IDeleteListService, ListRequestDto } from "@protocols";
 
 export class DeleteListService implements IDeleteListService {
@@ -9,7 +10,10 @@ export class DeleteListService implements IDeleteListService {
   constructor(private clientRepository: IClientRepository) { }
 
   public async delete(request: ListRequestDto): Promise<void> {
-    await this.clientRepository.get(this.params(request));
+    const data = await this.clientRepository.get(this.params(request));
+    if (!data.Item) {
+      throw new NotFoundError('To-do list not found with given identifier');
+    }
     await this.clientRepository.delete(this.params(request));
     const tasks = await this.clientRepository.query(this.queryTasksParams(request));
     await this.deleteTasks(tasks);
